move concatdataset to common
