+++ nemo/collections/common/__init__.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import nemo.collections.common.callbacks
-from nemo.collections.common import losses, parts, tokenizers
+from nemo.collections.common import data, losses, parts, tokenizers
 from nemo.package_info import __version__
 
 # Set collection version equal to NeMo version.
