+++ nemo/collections/nlp/data/__init__.py
@@ -27,7 +27,6 @@ from nemo.collections.nlp.data.language_modeling.lm_bert_dataset import (
 )
 from nemo.collections.nlp.data.language_modeling.sentence_dataset import SentenceDataset, TarredSentenceDataset
 from nemo.collections.nlp.data.machine_translation.machine_translation_dataset import (
-    ConcatTranslationDataset,
     TarredTranslationDataset,
     TranslationDataset,
 )
