+++ nemo/collections/nlp/models/machine_translation/mt_enc_dec_model.py
@@ -28,13 +28,14 @@ from pytorch_lightning import Trainer
 from pytorch_lightning.utilities import rank_zero_only
 from sacrebleu import corpus_bleu
 
+from nemo.collections.common.data import ConcatDataset
 from nemo.collections.common.losses import NLLLoss, SmoothedCrossEntropyLoss
 from nemo.collections.common.metrics import GlobalAverageLossMetric
 from nemo.collections.common.parts import transformer_weights_init
 from nemo.collections.common.tokenizers.chinese_tokenizers import ChineseProcessor
 from nemo.collections.common.tokenizers.en_ja_tokenizers import EnJaProcessor
 from nemo.collections.common.tokenizers.moses_tokenizers import MosesProcessor
-from nemo.collections.nlp.data import ConcatTranslationDataset, TarredTranslationDataset, TranslationDataset
+from nemo.collections.nlp.data import TarredTranslationDataset, TranslationDataset
 from nemo.collections.nlp.models.enc_dec_nlp_model import EncDecNLPModel
 from nemo.collections.nlp.models.machine_translation.mt_enc_dec_config import MTEncDecModelConfig
 from nemo.collections.nlp.modules.common import TokenClassifier
@@ -470,7 +471,7 @@ class MTEncDecModel(EncDecNLPModel):
                     datasets.append(dataset)
 
                 if self.multilingual:
-                    dataset = ConcatTranslationDataset(
+                    dataset = ConcatDataset(
                         datasets=datasets,
                         sampling_technique=cfg.get('concat_sampling_technique'),
                         sampling_temperature=cfg.get('concat_sampling_temperature'),
@@ -522,7 +523,7 @@ class MTEncDecModel(EncDecNLPModel):
                 datasets.append(dataset)
 
             if self.multilingual:
-                dataset = ConcatTranslationDataset(
+                dataset = ConcatDataset(
                     datasets=datasets,
                     shuffle=cfg.get('shuffle'),
                     sampling_technique=cfg.get('concat_sampling_technique'),
