+++ nemo/collections/tts/models/talknet.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 from collections import OrderedDict
-
 from typing import List
 
 import torch
